fix(checker): satisfy class heritage lint guards
